--- dbsql/data.csv
@@ -1,186 +1,0 @@
-UserID::FoodID
-1::1014
-1::5516
-1::8923
-1::6431
-1::1568
-1::1896
-1::2689
-1::4313
-1::5123
-1::5133
-1::6855
-1::9566
-1::10151
-2::1569
-2::4236
-2::9005
-2::6426
-2::8912
-2::4334
-2::5656
-2::1024
-2::1189
-2::3520
-2::6461
-2::7515
-2::8654
-2::8775
-2::9016
-2::9089
-2::10516
-3::1005
-3::1024
-3::1568
-3::2563
-3::2897
-3::3569
-3::4869
-3::5178
-3::6218
-3::7844
-3::8848
-3::9159
-3::10111
-4::4265
-4::8901
-4::1568
-4::1888
-4::2074
-4::3555
-4::4120
-4::5109
-4::5304
-4::6684
-4::7862
-4::8868
-4::9159
-4::10658
-5::7742
-5::2542
-5::3843
-5::9575
-5::1568
-5::1917
-5::2564
-5::3513
-5::3815
-5::4520
-5::6590
-5::8868
-5::9566
-5::10156
-6::4752
-6::2572
-6::1634
-6::5686
-6::1068
-6::1865
-6::2099
-6::3513
-6::4596
-6::5866
-6::9516
-6::10189
-7::3656
-7::8567
-7::1545
-7::3678
-7::7145
-7::9973
-7::1994
-7::1081
-7::1958
-7::3646
-7::5689
-7::6959
-7::7888
-7::8000
-7::9155
-7::10006
-8::5456
-8::1745
-8::2365
-8::3745
-8::6665
-8::1895
-8::3964
-8::7597
-8::1081
-8::3655
-8::5445
-8::6321
-8::7531
-8::8524
-8::9517
-8::10025
-9::6644
-9::1745
-9::3745
-9::5465
-9::1564
-9::2584
-9::3961
-9::4890
-9::1001
-9::2563
-9::3548
-9::3984
-9::4856
-9::5694
-9::6454
-9::7592
-9::8624
-9::9542
-9::10000
-10::7004
-10::9015
-10::5250
-10::9854
-10::1535
-10::2544
-10::3554
-10::4525
-10::5951
-10::6454
-10::7895
-10::8542
-10::9541
-10::10111
-10::11056
-11::1504
-11::6305
-11::8589
-11::9901
-11::1584
-11::1695
-11::2997
-11::1099
-11::1198
-11::1699
-11::1759
-11::2599
-11::3105
-11::3699
-11::4566
-11::4988
-11::5466
-11::5987
-11::6428
-11::7534
-11::8624
-11::10111
-12::1567
-12::2564
-12::6358
-12::1896
-12::2689
-12::2515
-12::6984
-1::10
-1::11
-1::1014
-1::5516
-1::12
-1::34
-1::56
