--- server/123456.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpgIBAAKCAQEAykJQOgOj0RmBC9TBp7+hUkl2cvK5VCK/QYJi+DxRakRkkd6P
-rCCSy6Fl0IQ2f6qNsEPkWj86Ddabc/vZXpLqw+da18ldk/xHkT0f/BdEo2O8N4IE
-mHZs+VYICHiccJAkSwP+JSbjiRCpaLoVJYnw9JDA8K71ulqW8mx4eAZBgaf5A5vK
-cRJuzgzUx3I0iK9HMgkimeeIf9a4CokE2D/4ze9hMXu8uyAa2v6PsYsPY5csMIxN
-br176fcWiyyFX6Jw1u9yxy26PDOqXkBe+pqsVWCxnfdjp9C79kTzgHj0THiRxJoo
-jHzSPsGRyG/BlL9mhPhkfmdI5q8GSwSbdDKO4QIDAQABAoIBAQCgLQYV+533UYhW
-E0bODz+1yAHrJHhKvj3Spmy0hChBzSoML0J+oUk/xHC6QopabnpJNmsYYWwbW3K+
-1pudfyI2xPi3XXwjXy5aQhsbbF3qO2jqWKhFCbxhdZVlCAd0SWLEmcPiLjsR6EUR
-cBShJu3Q8PYrZZOq4mUaVXbOLr4f67l031TYeExspli+rBLyMihXQN5OuyMZzOGi
-jZnNrlnwcRn3q9NPuUohYZdLPy5FxNbYYoIjnoUD+wQs5uaj+vlliCaMzioj7GnX
-zLC/izcOeCt4muBuW/XdRdeiJbugRnt1XbsD6SoIQaIbWrmnUetzMgbDgcdTAhjE
-n8NJNWnBAoGBAPrQRINsJs4TrZaGxPKiBZEjAqAbeZ7AsC+AbzA0p38L2MDDLHOU
-sguBcx6cdZpCL0XXbjKAqRigDQ1dudKzfpiER6B/imKMTac0/XhNbKRuHPDxvotH
-rGZ14NPl00Id/5iigebesn+x+A6w7mIclPv0AQipl9XIdvvf6Rp36xlZAoGBAM5x
-AzrEmGwtgb+veIQ09JplOCHtkMWK7PbhrPqqvzZQB6FFQf9tP1DQq4s4Zd+57pe0
-KMqQjorOIPuWx89XK5jUZDyAMLFI8f6VMiXQ20VeUEbF2XS/l9YjniZDUi9+6mbJ
-oUiXTWSo7veqcwd+YfjHUxpSEBDavuAwH1OzW+jJAoGBAKjpKbr4Uwnl5rnA+VIO
-G4Gq+lGpLzbivcS/0iG81DCTYd4vkvgj4+QrwcOTyv+gIwX+F5CbZUXYQ/IEdOiZ
-SWCddT95pVDI/GLnJOyn98qk/YimxZQ3KLu4FzmdRbD5UvEQQDujLsMgr9X1AhXA
-+fTUVv3meQLYhokKOopAyGvBAoGBAMZ9qPJHdpqEwr0wlPrhA9Kzwd31Un5tjlvu
-LuixZbpuBngMZGGYhAlLKbVpm6IaXwrkAViK11Mk0Fe26RU1pkiC5oG5a5f6IBt9
-DkOH5lxzS1JRBz2WvgH6KyR1FD9wkyYTdgya1aIUP/hHneQ1ozOVSxUClEnwWRir
-jGNpD9TpAoGBAMu4f2xBfD5Xf2u3DEL/eITWhxSHqeR5hC4Sp3VhBoVLQTiQHsm9
-HNTFIkhf4QU+QsB5MikoFGy5NZIhR3D9xLvTJkRL04ECUlaTqNuUB7zZkBxpsXob
-uNSpoMAaqbwasQfAZ+QqFxnWN64CI4RduDtxjw+4UAOXWmjjXZLyicAg
------END RSA PRIVATE KEY-----
